refactor: tests folder structure
